Move SamplingConfig out of trace config
