+++ agent/agent-plugins/apache-kafka/src/main/java/org/bithon/agent/plugin/apache/kafka/consumer/interceptor/ListenerConsumer$PollAndInvoke.java
@@ -23,7 +23,7 @@ import org.bithon.agent.bootstrap.aop.IBithonObject;
 import org.bithon.agent.bootstrap.aop.InterceptionDecision;
 import org.bithon.agent.core.config.ConfigurationManager;
 import org.bithon.agent.core.tracing.Tracer;
-import org.bithon.agent.core.tracing.config.TraceConfig;
+import org.bithon.agent.core.tracing.config.TraceSamplingConfig;
 import org.bithon.agent.core.tracing.context.ITraceContext;
 import org.bithon.agent.core.tracing.context.ITraceSpan;
 import org.bithon.agent.core.tracing.context.TraceContextFactory;
@@ -51,7 +51,7 @@ public class ListenerConsumer$PollAndInvoke extends AbstractInterceptor {
 
     private final ISampler sampler = SamplerFactory.createSampler(ConfigurationManager.getInstance()
                                                                                       .getDynamicConfig("tracing.samplingConfigs.kafka-consumer",
-                                                                                                        TraceConfig.SamplingConfig.class));
+                                                                                                        TraceSamplingConfig.class));
 
     /**
      * Keep topic information on injected fields for further use
